# services/cms/docs/openapi.yaml
openapi: 3.0.3
info:
  title: BusinessLogic CMS API
  description: |
    Complete API for the BusinessLogic CMS platform.

    **Audience markers:**
    - `x-internal: true` — frontend-only endpoints (called by Directus extension UI)
    - `x-formula-api: true` — callback endpoints (called BY the Formula API service)
    - Neither — public endpoints (registration, Stripe webhooks)

    **Excluded:** 3 formula execution proxy routes (`/calc/formula/execute*`) — see Formula API docs.
  version: 1.0.0
servers:
  - url: '{cmsBaseUrl}'
    description: Directus CMS instance
    variables:
      cmsBaseUrl:
        default: http://localhost:8055
security:
  - bearerAuth: []
...
paths:
...
  /calc/admin/overview:
    get:
      operationId: adminOverview
      summary: Admin KPI overview
      description: Returns aggregated KPIs (accounts, subscriptions, calculators, calls, revenue) and chart data.
      tags: [Admin Dashboard]
      x-internal: true
      responses:
        '200':
          description: Overview KPIs and charts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOverview'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /calc/admin/accounts:
    get:
      operationId: adminListAccounts
      summary: Paginated account list
      description: Lists accounts with subscription status, calculator counts, and monthly calls.
      tags: [Admin Dashboard]
      x-internal: true
      parameters:
        - $ref: '#/components/parameters/searchParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [trialing, active, canceled, expired]
        - $ref: '#/components/parameters/paginationPage'
        - $ref: '#/components/parameters/paginationLimit'
        - $ref: '#/components/parameters/sortParam'
      responses:
        '200':
          description: Account list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminAccountSummary'
                  meta:
                    $ref: '#/components/schemas/PaginatedMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /calc/admin/accounts/{accountId}:
    get:
      operationId: adminAccountDetail
      summary: Account detail
      description: Returns account info, subscription, calculators with profiles, and 30-day usage chart.
      tags: [Admin Dashboard]
      x-internal: true
      parameters:
        - $ref: '#/components/parameters/accountIdPath'
      responses:
        '200':
          description: Account detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAccountDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /calc/admin/calculators:
    get:
      operationId: adminListCalculators
      summary: Cross-account calculator list
      description: Lists all calculators with account info, profile, monthly calls, and config metadata.
      tags: [Admin Dashboard]
      x-internal: true
      parameters:
        - $ref: '#/components/parameters/searchParam'
        - $ref: '#/components/parameters/paginationPage'
        - $ref: '#/components/parameters/paginationLimit'
        - $ref: '#/components/parameters/sortParam'
      responses:
        '200':
          description: Calculator list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminCalculatorSummary'
                  meta:
                    $ref: '#/components/schemas/PaginatedMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /calc/admin/calculators/{calcId}/errors:
    get:
      operationId: adminCalculatorErrors
      summary: Recent calculator errors
      tags: [Admin Dashboard]
      x-internal: true
      parameters:
        - $ref: '#/components/parameters/calcIdPath'
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Error list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        timestamp:
                          type: string
                          format: date-time
                        error_message:
                          type: string
                          nullable: true
                        response_time_ms:
                          type: integer
                        test:
                          type: boolean
                        cached:
                          type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /calc/admin/extend-trial:
    post:
      operationId: adminExtendTrial
      summary: Extend account trial
      tags: [Admin Dashboard]
      x-internal: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accountId, days]
              properties:
                accountId:
                  type: string
                  format: uuid
                days:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Trial extended
          content:
            application/json:
              schema:
                type: object
                properties:
                  trial_end:
                    type: string
                    format: date-time
        '400':
          description: Missing fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /calc/admin/set-exempt:
    post:
      operationId: adminSetExempt
      summary: Toggle subscription exemption
      tags: [Admin Dashboard]
      x-internal: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accountId, exempt]
              properties:
                accountId:
                  type: string
                  format: uuid
                exempt:
                  type: boolean
      responses:
        '200':
          description: Exemption updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  exempt:
                    type: boolean
        '400':
          description: Missing fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ──────────────────────────────────────────────────────────
  # Admin Infrastructure
  # ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Directus static token (Formula API role) or user access token
  parameters:
    calcIdPath:
      name: calcId
      in: path
      required: true
      description: Calculator ID (user-defined string, e.g. `my-calculator` or `my-calculator-test`)
      schema:
        type: string
    accountIdPath:
      name: accountId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    paginationPage:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    paginationLimit:
      name: limit
      in: query
      schema:
        type: integer
        default: 25
        maximum: 100
    sortParam:
      name: sort
      in: query
      description: 'Sort field. Prefix with `-` for descending (e.g. `-date_created`)'
      schema:
        type: string
        default: '-date_created'
    searchParam:
      name: search
      in: query
      schema:
        type: string
  schemas:
...
    AdminOverview:
      type: object
      properties:
        accounts:
          type: object
          properties:
            total:
              type: integer
        subscriptions:
          type: object
          properties:
            by_plan:
              type: array
              items:
                type: object
                properties:
                  plan:
                    type: string
                  count:
                    type: integer
        calculators:
          type: object
          properties:
            total:
              type: integer
            active:
              type: integer
        calls:
          type: object
          properties:
            today:
              type: integer
            week:
              type: integer
            month:
              type: integer
            errors_month:
              type: integer
        revenue:
          type: object
          properties:
            mrr:
              type: integer
            active_subscriptions:
              type: integer
            churned_30d:
              type: integer
            trial_total:
              type: integer
            trial_converted:
              type: integer
        charts:
          type: object
          properties:
            calls_per_day:
              type: array
              items:
                type: object
                properties:
                  date:
                    type: string
                    format: date
                  total:
                    type: integer
                  errors:
                    type: integer
            accounts_per_month:
              type: array
              items:
                type: object
                properties:
                  month:
                    type: string
                  count:
                    type: integer
            deletions_per_month:
              type: array
              items:
                type: object
                properties:
                  month:
                    type: string
                  count:
                    type: integer
            conversions_per_month:
              type: array
              items:
                type: object
                properties:
                  month:
                    type: string
                  count:
                    type: integer
    AdminAccountSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        date_created:
          type: string
          format: date-time
        exempt_from_subscription:
          type: boolean
        subscription_status:
          type: string
          nullable: true
        trial_end:
          type: string
          format: date-time
          nullable: true
        plan_name:
          type: string
          nullable: true
        calculator_count:
          type: integer
        active_count:
          type: integer
        monthly_calls:
          type: integer
    AdminAccountDetail:
      type: object
      properties:
        account:
          type: object
        subscription:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid
            status:
              type: string
            plan_name:
              type: string
            calculator_limit:
              type: integer
              nullable: true
            calls_per_month:
              type: integer
              nullable: true
            calls_per_second:
              type: integer
              nullable: true
            trial_end:
              type: string
              format: date-time
              nullable: true
        calculators:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              activated:
                type: boolean
              over_limit:
                type: boolean
              date_created:
                type: string
                format: date-time
              date_updated:
                type: string
                format: date-time
                nullable: true
              profile:
                type: object
                nullable: true
              monthly_calls:
                type: integer
              monthly_errors:
                type: integer
        usage:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              total:
                type: integer
              errors:
                type: integer
    AdminCalculatorSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        activated:
          type: boolean
        over_limit:
          type: boolean
        date_created:
          type: string
          format: date-time
        date_updated:
          type: string
          format: date-time
          nullable: true
        account_name:
          type: string
        account_id:
          type: string
          format: uuid
        monthly_calls:
          type: integer
        monthly_errors:
          type: integer
        profile:
          type: object
          nullable: true
        unresolved_functions:
          type: array
          nullable: true
          items:
            type: object
            properties:
              name:
                type: string
              references:
                type: array
                items:
                  type: string
        config_version:
          type: integer
          nullable: true
        file_version:
          type: integer
          nullable: true
...
    PaginatedMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            errors:
              - message: Authentication required
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            errors:
              - message: Access denied
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            errors:
              - message: Not found